Podcast audio can have larger content-length.

+++ rinse_rss/rinsefm/models.py
@@ -6,7 +6,7 @@ class PodcastEpisode(models.Model):
     title = models.CharField(max_length=200)
     broadcast_date = models.DateTimeField()
     audio_url = models.URLField(unique=True)
-    audio_content_length = models.PositiveSmallIntegerField()
+    audio_content_length = models.PositiveIntegerField()
     audio_content_type = models.CharField(max_length=200)
 
     def __str__(self):
